move invariants out of simulation packages
